Rename meta kv processors
