+++ src/meta/MetaServiceHandler.cpp
@@ -33,12 +33,12 @@
 #include "meta/processors/index/ListEdgeIndexesProcessor.h"
 #include "meta/processors/index/FTServiceProcessor.h"
 #include "meta/processors/index/FTIndexProcessor.h"
-#include "meta/processors/customKV/MultiPutProcessor.h"
-#include "meta/processors/customKV/GetProcessor.h"
-#include "meta/processors/customKV/MultiGetProcessor.h"
-#include "meta/processors/customKV/ScanProcessor.h"
-#include "meta/processors/customKV/RemoveProcessor.h"
-#include "meta/processors/customKV/RemoveRangeProcessor.h"
+#include "meta/processors/kv/MultiPutProcessor.h"
+#include "meta/processors/kv/GetProcessor.h"
+#include "meta/processors/kv/MultiGetProcessor.h"
+#include "meta/processors/kv/ScanProcessor.h"
+#include "meta/processors/kv/RemoveProcessor.h"
+#include "meta/processors/kv/RemoveRangeProcessor.h"
 #include "meta/processors/admin/HBProcessor.h"
 #include "meta/processors/user/AuthenticationProcessor.h"
 #include "meta/processors/admin/BalanceProcessor.h"
