+++ src/meta/test/ProcessorTest.cpp
@@ -24,12 +24,12 @@
 #include "meta/processors/schema/ListEdgesProcessor.h"
 #include "meta/processors/schema/AlterTagProcessor.h"
 #include "meta/processors/schema/AlterEdgeProcessor.h"
-#include "meta/processors/customKV/MultiPutProcessor.h"
-#include "meta/processors/customKV/GetProcessor.h"
-#include "meta/processors/customKV/MultiGetProcessor.h"
-#include "meta/processors/customKV/RemoveProcessor.h"
-#include "meta/processors/customKV/RemoveRangeProcessor.h"
-#include "meta/processors/customKV/ScanProcessor.h"
+#include "meta/processors/kv/MultiPutProcessor.h"
+#include "meta/processors/kv/GetProcessor.h"
+#include "meta/processors/kv/MultiGetProcessor.h"
+#include "meta/processors/kv/RemoveProcessor.h"
+#include "meta/processors/kv/RemoveRangeProcessor.h"
+#include "meta/processors/kv/ScanProcessor.h"
 #include "meta/processors/zone/AddGroupProcessor.h"
 #include "meta/processors/zone/DropGroupProcessor.h"
 #include "meta/processors/zone/ListGroupsProcessor.h"
